NLP in Python
lemma, stem words and bag of words. Three of the most of important characteristics ever found in the literature of Natural Language Processing. In order to run the train.py file you shall create a conda environment into the linux terminal through the bash environment in the first place. Furthermore, you shall find much more information in the Anaconda file in the repository.